pyproject.toml
[project]
name = "book"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.21.0",
    "asyncpg>=0.30.0",
    "bcrypt<4.0",
    "beautifulsoup4>=4.13.4",
    "cython>=3.1.0",
    "dotenv>=0.9.9",
    "fastapi[standard]>=0.115.12",
    "lxml>=5.4.0",
    "mobi>=0.3.3",
    "passlib[bcrypt]>=1.7.4",
    "pathlib>=1.0.1",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.4",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "requests>=2.32.3",
    "sqlalchemy>=2.0.40",
    "uvicorn>=0.34.2",
]
